rename folder myrequest as dayoff, file myrequest as dayoff
